Use a specific zlib version
tests\data\conda_format_repo\win-64 subdir has multiple zlib versions in repodata.json. Since the index is a dict, python 2.7 tests might see either zlib version first.  One versoin of zlib (1.2.11) has both a bzip and conda package, while the other (1.2.8) has only a bzip. The point of the test is to ensure conda packages are preferred over bzip, so the test always fails if zlib 1.2.8 happens to be the first key found while iterating the index. Python 3.7 preserves insertion order, so the tests is stable in that version.